requirements.txt
numpy
scipy
scikit-learn
matplotlib
seaborn
torcheval
pandas
# torch